+++ app/class-core.php
@@ -180,6 +180,7 @@ class Core {
 	 * @see Integrations\Shortpixel
 	 * @see Integrations\JS_Composer
 	 * @see Integrations\Flatsome
+	 * @see Integrations\AIO_SEO
 	 */
 	private function init_integrations() {
 		$loader = Loader::get_instance();
@@ -193,6 +194,7 @@ class Core {
 		$loader->integration( 'elementor' );
 		$loader->integration( 'js-composer' );
 		$loader->integration( 'flatsome' );
+		$loader->integration( 'aio-seo' );
 	}
 
 	/**