+++ assets/_src/app.tsx
@@ -11,12 +11,13 @@ import './app.scss';
 import Nav from './components/nav';
 import SettingsProvider from './context/provider';
 import CloudflareSettings from './routes/cloudflare/settings';
-import CloudflareExperimental from './routes/cloudflare/experimental';
+import Experimental from './routes/cloudflare/experimental';
 import Support from './routes/support';
 import ToolsSettings from './routes/tools/settings';
 import Logs from './routes/misc/logs';
 import ImageGenerateRoute from './routes/image/generate';
 import ToolsPremium from './routes/tools/premium';
+import Integrations from './routes/cloudflare/integrations';
 
 /**
  * App
@@ -24,6 +25,25 @@ import ToolsPremium from './routes/tools/premium';
  * @class
  */
 const App = () => {
+	const routes = () => {
+		return (
+			<Routes>
+				<Route index element={<CloudflareSettings />} />
+				<Route path="/cf/integrations" element={<Integrations />} />
+				<Route path="/cf/experimental" element={<Experimental />} />
+				<Route path="/tools/settings" element={<ToolsSettings />} />
+				<Route path="/tools/premium" element={<ToolsPremium />} />
+				<Route
+					path="/image/generate"
+					element={<ImageGenerateRoute />}
+				/>
+				<Route path="/misc/logs" element={<Logs />} />
+				<Route path="/misc/support" element={<Support />} />
+				<Route path="*" element={<Navigate to="/" replace />} />
+			</Routes>
+		);
+	};
+
 	return (
 		<HashRouter>
 			<SettingsProvider>
@@ -32,33 +52,7 @@ const App = () => {
 				</div>
 
 				<div className="column">
-					<div className="box">
-						<Routes>
-							<Route index element={<CloudflareSettings />} />
-							<Route
-								path="/cf/experimental"
-								element={<CloudflareExperimental />}
-							/>
-							<Route
-								path="/tools/settings"
-								element={<ToolsSettings />}
-							/>
-							<Route
-								path="/tools/premium"
-								element={<ToolsPremium />}
-							/>
-							<Route
-								path="/image/generate"
-								element={<ImageGenerateRoute />}
-							/>
-							<Route path="/misc/logs" element={<Logs />} />
-							<Route path="/misc/support" element={<Support />} />
-							<Route
-								path="*"
-								element={<Navigate to="/" replace />}
-							/>
-						</Routes>
-					</div>
+					<div className="box">{routes()}</div>
 				</div>
 			</SettingsProvider>
 		</HashRouter>