+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Added:
 * Integration with Smart Slider 3
+* Integration with All in One SEO: allow controlling application/ld+json schema image URLs
 * Filter cf_images_disable_crop to disable auto cropping for registered crop images
 
 Changed: